feat: send `OPEN_TAB` action to browser with user selected tab to open
I am so so disheartened with Emacs's builtin completing-read scene. It just
can't fathom that more than a list of plain strings might need to be completed.
I gave up on making it do what I wanted, and clobbered everything into the
user-visible string. May be in future when I know of a way to make Emacs:

1. Show only the title in selection prompt, with URL as annotation
2. Perform search on both title and URL
3. Provide me all params of the tab, not just the string which was shown to user

I'll revisit how the completion is done, but for now I suppose I just have to
live with this ugly UI.