refactor: major architectural shift in spookfox-addon
Also added documentation wherever I could. I believe things are more declarative
now, and easier to understand. It should now be easier to work with Emacs side
of things without the knowledge of how spookfox-native is involved; just do a
`spookfox.request` and get whatever Emacs responded with.

State management is also more predictable and declarative now. All the UI can be
driven with a subscription to `SFEvents.NEW_STATE` event. Next I need to figure
out a clean way of doing UI.

Apologies @msanfilippof for changing things so dramatically while you were
probably working on the same code. #9 is still open because it needs an overall
refactoring and a finer comb to make sure everything is consistent.