feat: reliably working chained tab sync with firefox
All the changes made to a tab (opening, visiting different URLs, closing) are
now reliably (in my testing so far, which isn't much) working.

Tabs are chained by default, i.e when a tab is opened, we store it in Emacs.
Only way for user to delete the tab from Emacs is to do it by hand, or close the
tab *while* it is in chained mode.

I think we should get rid of 'TOGGLE_TAB_CHAINING' request from browser, and
instead use 'UPDATE_TAB' for that. This works better for scenarios like: User
opened a tab, it is chained at this point. User unchains the tab. User go to Y
URL. User chains the tab again. Now even though tab chaining is set to `t` in
Emacs, URL and title for this tab are still old.