refactor: redux style state management
So we don't end up with hairy race-conditions during state changes.
Fixes #17

With this, I think I have a way forward for extensibility I wanted. I want
spookfox to be extensible from Emacs itself. i.e any time we want to do
something in Browser, we should be able to do it without rebuilding the Firefox
addon. We can have spookfox itself as a core which takes care of communication
b/w Emacs and browser, and provides an abstraction for state-management (i.e
maintaining some data in browser for situations like tab-management). And then
things like tab-management can be apps(?), which extend spookfox using
spookfox-core. I will try implementing tab-management this way to see if it is
as feasible as I think it is (i.e #21).