feat: new "apps" architecture
Apps are plugins or extensions which use spookfox as a core to add actual
features, like managing tabs using org-mode (JS side of which is implemented in
this commit). I chose to call them "apps" instead of "plugin" or "extension",
because spookfox itself is a browser extension/plugin, and having extensions of
an extension might be confusing.

This commit implements a working version of tab-management as an app, but it
needs to be polished and refactored.

Fix #21