fix: gobble errors of event-handlers
In the addon, Firefox sometimes don't report errors that occur in async
event-listeners (or in event-listeners, async or not, I didn't verify). This
commit creates a helper function which wraps the event-handler in a try-catch
block, and just logs the error instead of throwing it. I believe this was the
reason of #14; but I am not sure. We'll see.

Also,
- refactor: make better use of immer in OrgTabs.reducer to simplify state updates