fix: remove unnecessary redirection done for browser->socket-clients
For some reason, this also make #14 un-reproducible. I was able to reproduce it
reliably by sending 277 'GET_SAVED_TABS' messages from Firefox in a loop. But
after making this change, I can't do that anymore.

I had built the understanding that the bug was caused by `conn.write` call in
`rx` loop, which would just hang up and not write anymore. I don't know why, but
that is not happening anymore.

Maybe this means the bug is resolved? I have thought the same more than once
earlier as well. Let's put this baby to test for a few days and then we'll see
if it actually get fixed, or if...urgh...I don't want to feel that pain again.