refactor: split spookfox-tabs and spookfox-org-tabs
Saving tabs in org file and general-purpose access to browser tabs should be
separate.

I have kinda make a mess in there with =msg-prefix=. It should be refactored
properly; there should not be a need to set the prefix on both emacs and TS
side. Perhaps when ENABLE_APP message is sent to browser, it should also carry
the prefix which will be used for this app's messages.