fix: #26
This commit fixes the leak which caused indefinite number of auto-connect
intervals getting created, and also fixes the memory leak which was caused due
to directly accessing the background page in the addon popup. Directly accessing
the background-page in addon's popup lead to the Spookfox object getting cloned
every time the popup is opened. Even when GC kills the object in background
script, its clones stayed alive and were causing memory leaks.