fix: add content-security-policy to allow injecting js in background
I am feeling so stupid. What makes it worse is that I actually did search how to
set CSP for a web-extension but somehow didn't find a way. I found it now
because `parcel serve` is setting the CSP to the value I've added now, but don't
do it for `parcel build`; which adds to the list of reasons for getting rid of
Parcel.