feat: spookfox-js-injection-eval to inject js in background script
Since it is possible to run js in background script, why not just do that! Well,
one reason is that it is kinda ugly to write js in lisp, but we can always write
js in a separate file and then inject its contents.

Running CL using jscl has a cost of ~500ms per-call. Eval-ing js takes 3ms,
eval-ing CL takes 300. I guess it is good to have an option at least; we can use
jscl for one-time-setup to create JS functions, and then for interactive
operations, we can eval js instead.