fix: autoloads and fix package formatting
turns out, installing package using `package-install-from-file' was failing because of some silly
things in spookfox.el's header. I still couldn't use package.el in my own config though, it was
failing with error "package websocket not available". I assume it is because of Elpaca.

I will package spookfox for guix soon and use that. Both straight.el and Elpaca turned out to be a
pain in the ass after a while.